resume_analyzer/requirements.txt
Django==4.2
PyPDF2==3.0.0
python-docx==0.8.11
requests==2.31.0
